refactor(e2e): split harness, unify helpers, add Makefile, drop CI dupes
Self-review cleanup of the e2e suite:

- Split the 459-line harness_test.go: run machinery stays in harness_test.go;
  assertions, output parsers, and backend-gap/skip helpers move to
  helpers_test.go.
- Unify the duplicated packageBackendGap/scanBackendGap into a single
  backendGap() helper; use the clearer errorLine() in their skip messages.
- Fix a temp-dir leak: the freshly built binary's dir is now removed on exit.
- Drop the write-only result.err field and the dead expectedName var.
- Remove the duplicated e2e/ci/ CI snippets; the canonical wiring lives in the
  JuliaHub repo. Fold a concise "Running in CI" section into e2e/README.md.
- Add a Makefile (make e2e / e2e-fast / e2e-compile / check) so developers
  don't have to remember the -tags e2e invocations; e2e-fast skips the slow
  project-list tests (~14s vs ~70s).
