+++ e2e/README.md
@@ -7,15 +7,16 @@ registry and its packages synced.
 
 The tests are **implemented here** (versioned alongside the CLI) but **executed
 by the JuliaHub platform nightly CI**, which deploys the instance and provisions
-credentials. See [`ci/`](./ci).
+credentials. See [Running in CI](#running-in-ci).
 
 ## Layout
 
-One file per command category, plus the shared harness:
+One file per command category, plus shared infrastructure:
 
 | File | Commands |
 |------|----------|
-| [harness_test.go](harness_test.go) | build/run helpers, isolated-config auth, output parsers, skip helpers |
+| [harness_test.go](harness_test.go) | `TestMain`, binary build, isolated-config auth, `runJH`/`runOK` |
+| [helpers_test.go](helpers_test.go) | assertions, output parsers, backend-gap/skip helpers |
 | [meta_test.go](meta_test.go) | `--version`, `--help` |
 | [auth_test.go](auth_test.go) | `auth status`, `auth env` |
 | [user_test.go](user_test.go) | `user info`, `user list`, identity cross-check |
@@ -48,24 +49,22 @@ One file per command category, plus the shared harness:
   response that fails to parse) are deliberately **not** matched by that helper,
   so they still fail.
 - **Admin commands** require elevated permissions: under a standard test user
-  they skip; run under an admin credential they execute. (See the note in
-  `ci/README.md` about credential selection.)
+  they skip; run under an admin credential they execute. (See
+  [Running in CI](#running-in-ci) about credential selection.)
 
 ## Design
 
 - **Build-tag gated.** Every file is tagged `//go:build e2e`, so the suite never
   runs as part of the normal `go test ./...` unit suite. Run it with
   `go test -tags e2e ./e2e/...`.
-- **GET / read-only.** The suite exercises only non-mutating commands, so it is
-  safe against a shared instance and needs no cleanup. (The CLI has no
-  `dataset delete`, so creating data would otherwise leave orphans.)
+- **Read-mostly.** The suite exercises read/GET commands, so it is safe against a
+  shared instance and needs no cleanup. The one exception is `scan`, which
+  submits a manifest scan with `--no-wait` (no long poll, nothing to delete).
+- **Adapts to the instance.** Tests assert real data/structure when the relevant
+  endpoint is available and **skip** when it is not (absent endpoint, missing
+  permissions, disallowed query, timeout). Genuine CLI defects still fail.
 - **Skips without credentials.** If no credentials are available, the suite
   skips rather than fails.
-
-> **Note on instance data.** The catalog tests assume the General registry and
-> common packages are synced — true for `nightly.juliahub.dev`. Pointed at a bare
-> instance (one with no registries synced), those tests will fail by design; the
-> identity and error-path groups still pass anywhere.
 
 ## Authentication
 
@@ -100,44 +99,52 @@ falls back to the real `~/.juliahub` (a developer's existing login).
 
 ## Running locally
 
-Against your existing login:
+Against your existing `jh auth login` (from the repo root):
 
 ```bash
+make e2e            # all tests
+make e2e-fast       # skips the slow project-list tests (GraphQL can be slow)
+
+# or directly:
 go test -tags e2e -v ./e2e/...
+go test -tags e2e -v -run TestDataset ./e2e/...     # one category
 ```
 
-With explicit credentials (matches the CI contract):
+Against a different instance without touching your login — supply explicit
+credentials and the harness writes a throwaway isolated config:
 
 ```bash
 JULIAHUB_SERVER=https://nightly.juliahub.dev \
-JULIAHUB_ID_TOKEN=<id_token> \
-JULIAHUB_TOKEN=<access_token> \
+JULIAHUB_ID_TOKEN="$(jh auth env | sed -n 's/^JULIAHUB_ID_TOKEN=//p')" \
   go test -tags e2e -v ./e2e/...
 ```
 
-To test a prebuilt binary instead of building from source, set `JH_BIN`:
+To test a prebuilt binary instead of rebuilding, set `JH_BIN`:
 
 ```bash
 JH_BIN=./jh go test -tags e2e -v ./e2e/...
 ```
 
-## Running from the JuliaHub platform nightly CI
+## Running in CI
 
-The platform CI already has everything needed:
+These tests run from the **JuliaHub platform** repo's nightly CI, which already
+deploys `nightly.juliahub.dev`, provisions test credentials, and has a headless
+login (`login-and-get-auth-toml`) that yields an `auth.toml`. The wiring there:
 
-1. A deployed nightly instance (`nightly.juliahub.dev`).
-2. Provisioned test credentials (`.github/workflows/reusable-get-credentials.yml`).
-3. A headless login that yields an `auth.toml`
-   (`.github/actions/login-and-get-auth-toml`).
+1. A composite action `.github/actions/jh-cli-e2e/` performs the headless login,
+   extracts the `id_token` **and** `access_token` from the `auth.toml`, checks
+   out this repo, and runs `go test -tags e2e ./e2e/...` with `JULIAHUB_SERVER` /
+   `JULIAHUB_ID_TOKEN` / `JULIAHUB_TOKEN` set (the harness turns those into a
+   `~/.juliahub`).
+2. A `jh-cli-e2e-tests` job in `platform-instance-tests.yml` (gated by a
+   `run-jh-cli-tests` input), modelled on the existing `juliahubjl-tests` job.
+3. `build-nightly.yml` enables it for the `nightly.juliahub.dev` run.
 
-To run this suite, add a job that checks out this repo, logs in to get an
-`auth.toml`, extracts the `id_token` (and `access_token`), and runs
-`go test -tags e2e`. A ready-to-use composite action and an example job are in
-[`ci/`](./ci):
-
-- [`ci/jh-cli-e2e/action.yml`](./ci/jh-cli-e2e/action.yml) — composite action to
-  copy into `JuliaHub/.github/actions/jh-cli-e2e/`.
-- [`ci/example-job.yml`](./ci/example-job.yml) — example job to add to
-  `platform-instance-tests.yml`, modeled on the existing `juliahubjl-tests` job.
-
-See [`ci/README.md`](./ci/README.md) for step-by-step integration instructions.
+> **Credential selection.** That job authenticates as the standard **test** user,
+> so the `admin *` tests skip there. To exercise them in CI, point the job at the
+> **admin** credential (already exposed by `reusable-get-credentials`); admin is a
+> superset, so the non-admin tests still pass.
+>
+> **Export the id token.** Most `jh` commands send the `id_token` as the bearer
+> token; only a few dataset operations use the access token. Make sure the CI
+> step exports `JULIAHUB_ID_TOKEN`, not only `JULIAHUB_TOKEN`.
