--- e2e/ci/example-job.yml
@@ -1,46 +1,0 @@
-# Example job to add to JuliaHub's .github/workflows/platform-instance-tests.yml.
-#
-# This mirrors the existing `juliahubjl-tests` job: it depends on
-# `get-credentials`, decrypts the shared test password, and invokes a composite
-# action that logs in and runs the suite. It runs `continue-on-error: true` so a
-# CLI regression is reported without failing the whole nightly pipeline (matching
-# the JuliaHub.jl job's behaviour); drop that once the suite is proven stable.
-#
-# Prerequisites:
-#   - Copy e2e/ci/jh-cli-e2e/ into JuliaHub/.github/actions/jh-cli-e2e/
-#   - Optionally gate behind a workflow input (e.g. inputs.run-jh-cli-tests),
-#     defined the same way as inputs.run-juliahubjl-tests.
-
-  jh-cli-e2e-tests:
-    runs-on: [self-hosted]
-    needs: [get-credentials, playwright-tests]
-    continue-on-error: true
-    outputs:
-      outcome: ${{ steps.outcome.outcome == 'success' && 'success' || 'failure' }}
-    timeout-minutes: 30
-    steps:
-      - name: Decrypt creds
-        id: decrypt
-        shell: bash
-        run: |
-          test_password_encrypted="${{ needs.get-credentials.outputs.test-password-encrypted }}"
-          secret="${{ secrets.GHA_OUTPUTS_ENCRYPTION_SECRET }}"
-          test_pass=$(gpg --decrypt --quiet --batch --passphrase "$secret" --output - <(echo "$test_password_encrypted" | base64 --decode))
-          echo "::add-mask::$test_pass"
-          echo "username=${{ needs.get-credentials.outputs.test-email }}" >> $GITHUB_OUTPUT
-          echo "password=$test_pass" >> $GITHUB_OUTPUT
-      - name: Pull down JuliaHub code
-        uses: actions/checkout@df4cb1c069e1874edd31b4311f1884172cec0e10 # v6
-      - name: jh CLI E2E tests
-        uses: ./.github/actions/jh-cli-e2e
-        # if: ${{ inputs.run-jh-cli-tests }}   # uncomment once a workflow input is added
-        with:
-          host: ${{ inputs.host }}
-          username: ${{ steps.decrypt.outputs.username }}
-          password: ${{ steps.decrypt.outputs.password }}
-          # Pin to a released tag once available; main is fine to start.
-          jh-ref: main
-      - id: outcome
-        name: Report Success 'outcome' Output
-        if: success()
-        run: echo "Job completed successfully"
